Refatorando o formulario de justificativas

+++ src/scielo/bin/xml/modules/xpmaker.py
@@ -1038,7 +1038,8 @@ def add_files_to_pmc_package(scielo_pkg_path, pmc_xml_filename, language):
     xml, e = xml_utils.load_xml(pmc_xml_filename)
     doc = article.Article(xml, xml_name)
     if language == 'en':
-        shutil.copyfile(scielo_pkg_path + '/' + xml_name + '.pdf', dest_path + '/' + xml_name + '.pdf')
+        if os.path.isfile(scielo_pkg_path + '/' + xml_name + '.pdf'):
+            shutil.copyfile(scielo_pkg_path + '/' + xml_name + '.pdf', dest_path + '/' + xml_name + '.pdf')
         for item in doc.href_files:
             shutil.copyfile(scielo_pkg_path + '/' + item.src, dest_path + '/' + item.src)
             validate_pmc_image(dest_path + '/' + item.src)
@@ -1107,25 +1108,26 @@ def pack_and_validate(xml_files, results_path, acron, version, is_db_generation=
         if not is_xml_generation:
             xpm_validations = pkg_reports.format_complete_report(report_components)
             filename = report_path + '/xml_package_maker.html'
+            if os.path.isfile(filename):
+                bkp_filename = report_path + '/xpm_bkp_' + '-'.join(utils.now()) + '.html'
+                shutil.copyfile(filename, bkp_filename)
             pkg_reports.save_report(filename, _('XML Package Maker Report'), xpm_validations.message, xpm_version())
+
             global DISPLAY_REPORT
             if DISPLAY_REPORT is True:
                 pkg_reports.display_report(filename)
 
-            header = html_reports.tag('p', _('Please, inform why each fatal error, error and warning is not valid.'))
-            header += html_reports.tag('p', _('It is very important and mandatory to complete this report and send it with the package.'))
-            header += html_reports.tag('p', _('It will be used to improve the messages and/or fix the validations made by XPM.'))
-            header += html_reports.tag('p', _('Complete the "{column}" column.').format(column=_('why it is not a valid message?')))
-            #header += html_reports.tag('p', _('Do not complete the "{column}" column.').format(column=_('answer')))
-
-            invalid_msg_report_filename = report_path + '/xpm_invalid_messages.html'
-            if os.path.isfile(invalid_msg_report_filename):
-                invalid_msg_report_bkp_filename = report_path + '/xpm_invalid_messages' + '-'.join(utils.now()) + '.html'
-                shutil.copyfile(invalid_msg_report_filename, invalid_msg_report_bkp_filename)
-
-            html_reports.save(invalid_msg_report_filename, _('XPM invalid messages'), header + pkg_reports.format_declaration_report(xpm_validations))
-            if DISPLAY_REPORT is True:
-                pkg_reports.display_report(invalid_msg_report_filename)
+            #header = html_reports.tag('p', _('Please, inform why each fatal error, error and warning is not valid.'))
+            #header += html_reports.tag('p', _('It is very important and mandatory to complete this report and send it with the package.'))
+            #header += html_reports.tag('p', _('It will be used to improve the messages and/or fix the validations made by XPM.'))
+            #header += html_reports.tag('p', _('Complete the "{column}" column.').format(column=_('why it is not a valid message?')))
+            #invalid_msg_report_filename = report_path + '/xpm_invalid_messages.html'
+            #if os.path.isfile(invalid_msg_report_filename):
+            #    invalid_msg_report_bkp_filename = report_path + '/xpm_invalid_messages' + '-'.join(utils.now()) + '.html'
+            #    shutil.copyfile(invalid_msg_report_filename, invalid_msg_report_bkp_filename)
+            #html_reports.save(invalid_msg_report_filename, _('XPM invalid messages'), header + pkg_reports.format_declaration_report(xpm_validations))
+            #if DISPLAY_REPORT is True:
+            #    pkg_reports.display_report(invalid_msg_report_filename)
 
         if not is_db_generation:
             if is_xml_generation:
